+++ src/volatility_formula.h
@@ -8,12 +8,12 @@
 
 
 // Function that returns the logarithmic return for a given time period
-std::vector<double> logarithmic_return_function(std::vector<double>& price) {
-    std::vector<double> r_t_list;
+std::vector<long double> logarithmic_return_function(std::vector<long double>& price) {
+    std::vector<long double> r_t_list;
     
     // Using a regular iterator to loop through all the initial prices
-    for (std::vector<double>::iterator it = price.begin(); it != price.end()-1; ++it) {
-        double r_t = log(*(it + 1) / *it);
+    for (std::vector<long double>::iterator it = price.begin(); it != price.end()-1; ++it) {
+        long double r_t = log(*(it + 1) / *it);
         r_t_list.push_back(r_t);
     
     }
@@ -23,25 +23,25 @@ std::vector<double> logarithmic_return_function(std::vector<double>& price) {
 
 // Function that returns the average of a vector of doubles
 // Return datatype: double 
-double average(std::vector<double>& return_list) {
-    double average = std::accumulate(return_list.begin(), return_list.end(), 0.0) / return_list.size();
+long double average(std::vector<long double>& return_list) {
+    long double average = std::accumulate(return_list.begin(), return_list.end(), 0.0) / return_list.size();
     return average;
 };
 
 
 // Function that returns the average return for a given time period
-double average_return(std::vector<double> r_average_list) {
-    double r_bar = average(r_average_list);
+long double average_return(std::vector<long double> r_average_list) {
+    long double r_bar = average(r_average_list);
     return r_bar;
 };
 
 
 // Function that returns a vector of the difference between a vector of 
 // log_return for that time period and the average return for the same time period
-double iter_variance(const std::vector<double>& log_return_for_time_period, double average_return_for_time_period) {
-        double variance_result = 0.0;
+long double iter_variance(const std::vector<long double>& log_return_for_time_period, long double average_return_for_time_period) {
+        long double variance_result = 0.0;
         int count = 0;
-        for (std::vector<double>::const_iterator it = log_return_for_time_period.begin(); it != log_return_for_time_period.end(); ++it) {
+        for (std::vector<long double>::const_iterator it = log_return_for_time_period.begin(); it != log_return_for_time_period.end(); ++it) {
             variance_result = std::pow((*it - average_return_for_time_period), 2.0) + variance_result;
             count++;
         };
@@ -52,29 +52,29 @@ double iter_variance(const std::vector<double>& log_return_for_time_period, doub
 
 
 // Function that returns the volatility for a given time period: Hourly in this case
-double volatility(std::vector<double>& log_return, double average) {
-    double variance = iter_variance(log_return, average)/(log_return.size() - 1);
-    double volatility = std::sqrt(variance);
+long double volatility(std::vector<long double>& log_return, long double average) {
+    long double variance = iter_variance(log_return, average)/(log_return.size() - 1);
+    long double volatility = std::sqrt(variance);
     return volatility;
 };
 
-
-double update_volatility(double old_volatility, double new_price, double old_price, double lambda) {
+// EWMA formula to update the volatility for a given time period 
+long double update_volatility(long double old_volatility, long double new_price, long double old_price, long double lambda) {
     // Calculate the log return
-    double r_t = log(new_price / old_price);
+    long double r_t = log(new_price / old_price);
     
     // Update variance using EWMA formula
-    double old_variance = std::pow(old_volatility, 2.0);
-    double new_variance = (1.0 - lambda) * std::pow(r_t, 2.0) + lambda * old_variance;
+    long double old_variance = std::pow(old_volatility, 2.0);
+    long double new_variance = (1.0 - lambda) * std::pow(r_t, 2.0) + lambda * old_variance;
     
     // Return the updated volatility (square root of variance)
     return std::sqrt(new_variance);
 };
 
-
-double volatility_algorithm(std::vector<double>& stock_prices) {
-        std::vector<double> log_returns = logarithmic_return_function(stock_prices);
-        double avg_return = average_return(log_returns);
-        double vol = volatility(log_returns, avg_return);
+// Function that returns the volatility for a given time period: Hourly for 6 consecutive hours
+long double volatility_algorithm(std::vector<long double>& stock_prices) {
+        std::vector<long double> log_returns = logarithmic_return_function(stock_prices);
+        long double avg_return = average_return(log_returns);
+        long double vol = volatility(log_returns, avg_return);
         return vol;
 };
